+++ Alpha_Engine/GSM.h
@@ -5,7 +5,7 @@ namespace GSM {
 	typedef void(*FP)(void);
 
 	extern int current, previous, next;
-	extern f64 gameTime;
+	extern f32 gameTime;
 	extern FP fpLoad, fpInit, fpUpdate, fpDraw, fpFree, fpUnload;
 
 	void GSM_Init(int startingState);
